params refactor
